scraper1/requirements.txt
argparse
beautifulsoup4
pyinstaller
requests
tqdm
python-dateutil
